Revert "WIP - more refactor fo IOW-356"
This reverts commit fe9fbe5810f41083604e5aff33256a56d0303a75.
