+++ src/main/java/gov/usgs/wma/waterdata/collections/BaseController.java
@@ -1,4 +1,4 @@
-package gov.usgs.wma.waterdata;
+package gov.usgs.wma.waterdata.collections;
 
 public abstract class BaseController {
 	protected String ogc404Payload = "{\"code\":\"404\", \"description\":\"Requested data not found\"}";
