+++ src/test/java/gov/usgs/wma/waterdata/collections/BoundingBoxTest.java
@@ -1,4 +1,4 @@
-package gov.usgs.wma.waterdata.features;
+package gov.usgs.wma.waterdata.collections;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
